Refactor: Move auth dependencies and simplify user db
Consolidate authentication dependencies into auth_backend.py and replace the old user_db.py with a simpler implementation.
