+++ backend/app/core/auth_backend.py
@@ -10,8 +10,10 @@ from fastapi_users.authentication.transport import BearerTransport
 from fastapi_users.password import PasswordHelper
 from beanie import PydanticObjectId
 from app.models.user import User, UserCreate
-from app.core.user_db import get_user_db
+from fastapi import Depends
+
 from app.core.config import settings
+from app.core.simple_user_db import get_simple_user_db
 import logging
 from typing import Optional
 
@@ -44,7 +46,7 @@ class UserManager(BaseUserManager[User, PydanticObjectId]):
 # UserManager dependency
 async def get_user_manager():
     """Get user manager instance"""
-    async for user_db in get_user_db():
+    async for user_db in get_simple_user_db():
         yield UserManager(user_db)
 
 # Bearer token transport
@@ -73,3 +75,15 @@ fastapi_users = FastAPIUsers[User, str](
 # Current user dependencies
 current_active_user = fastapi_users.current_user(active=True)
 current_superuser = fastapi_users.current_user(active=True, superuser=True)
+
+# Additional helper functions
+async def get_current_user_id(current_user: User = Depends(current_active_user)) -> str:
+    """Get current user ID as string"""
+    return str(current_user.id)
+
+async def get_current_user_optional() -> Optional[User]:
+    """Get current user if authenticated, otherwise None"""
+    try:
+        return await current_active_user()
+    except:
+        return None
